requirements.txt
Cython==0.29.7
influxdb==5.2.2
numpy==1.15.2
pandas==0.23.4
pyzmq==17.1.2